packages/python-scripts/requirements.txt
pinecone-client==3.0.2
python-dotenv==1.0.1
gensim==4.3.2
nltk==3.8.1 